requirements.txt
APScheduler==3.9.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
Django==3.2.8
django-braces==1.15.0
django-cleanup==5.2.0
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-filter==21.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
environ==1.0
mongojoin==1.0.0
pymongo==3.12.1
python-dotenv==0.19.0
requests==2.26.0
requests-file==1.5.1
schedule==1.1.0
mongoengine==0.20.0
sqlparse==0.2.4
